backend/requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
openpyxl  # for reading the excel files
scapy
mysql-connector-python
Flask
Flask-CORS